+++ Python/topoBatch/DemToTopoConsts.py
@@ -1,0 +1,11 @@
+COLOR_RELIEF_EXT = '_CR'
+HILL_SHADE_EXT = '_HS'
+SLOPE_EXT = '_SL'
+SLOPE_HILL_SHADE_EXT = '_SL_HS'
+SLOPE_WATER_EXT = '_SL_Water'
+SLOPE_POLY_EXT = '_SL_poly'
+TOPO_EXT = '_Topo'
+
+RED_VAL = 35
+GREEN_VAL = 170
+BLUE_VAL = 181
